Move review surface tests to review suite
