+++ Tests/QuillCodeAppTests/WorkspaceSurfaceTests.swift
@@ -740,137 +740,4 @@ final class WorkspaceSurfaceTests: XCTestCase {
         XCTAssertEqual(surface.archivedItems.first?.actions.map(\.kind), [.unarchive, .delete])
     }
 
-    func testGitDiffReviewSurfaceSummarizesLatestCompletedDiff() throws {
-        let diff = """
-        diff --git a/Sources/App.swift b/Sources/App.swift
-        index 1111111..2222222 100644
-        --- a/Sources/App.swift
-        +++ b/Sources/App.swift
-        @@ -1,3 +1,4 @@
-         import Foundation
-        -let title = "Old"
-        +let title = "QuillCode"
-        +let subtitle = "Review"
-        diff --git a/README.md b/README.md
-        index 3333333..4444444 100644
-        --- a/README.md
-        +++ b/README.md
-        @@ -1 +1 @@
-        -Old README
-        +New README
-        """
-        let call = ToolCall(name: "host.git.diff", argumentsJSON: "{}")
-        let result = ToolResult(ok: true, stdout: diff)
-        let thread = ChatThread(
-            title: "Review changes",
-            events: [
-                ThreadEvent(kind: .toolQueued, summary: "host.git.diff queued", payloadJSON: try JSONHelpers.encodePretty(call)),
-                ThreadEvent(kind: .toolRunning, summary: "host.git.diff running"),
-                ThreadEvent(kind: .toolCompleted, summary: "host.git.diff completed", payloadJSON: try JSONHelpers.encodePretty(result))
-            ]
-        )
-        let model = QuillCodeWorkspaceModel(root: QuillCodeRootState(
-            threads: [thread],
-            selectedThreadID: thread.id
-        ))
-
-        let review = model.surface().review
-
-        XCTAssertTrue(review.isVisible)
-        XCTAssertEqual(review.files.map(\.path), ["Sources/App.swift", "README.md"])
-        XCTAssertEqual(review.totalInsertions, 3)
-        XCTAssertEqual(review.totalDeletions, 2)
-        XCTAssertEqual(review.totalHunks, 2)
-        XCTAssertEqual(review.subtitle, "2 files changed, +3 -2")
-        XCTAssertEqual(review.files.first?.actions.map(\.kind), [.stage, .restore])
-        XCTAssertEqual(review.files.first?.hunkItems.count, 1)
-        XCTAssertEqual(review.files.first?.hunkItems.first?.actions.map(\.kind), [.stageHunk, .restoreHunk])
-        XCTAssertTrue(review.files.first?.hunkItems.first?.patch.contains("diff --git a/Sources/App.swift b/Sources/App.swift") == true)
-        let appLines = review.files.first?.hunkItems.first?.lines
-        XCTAssertEqual(appLines?.map(\.kind), [.context, .deletion, .insertion, .insertion])
-        XCTAssertEqual(appLines?.map(\.oldLineNumber), [1, 2, nil, nil])
-        XCTAssertEqual(appLines?.map(\.newLineNumber), [1, nil, 2, 3])
-    }
-
-    func testGitDiffReviewSurfaceIncludesMatchingReviewComments() throws {
-        let diff = """
-        diff --git a/Sources/App.swift b/Sources/App.swift
-        --- a/Sources/App.swift
-        +++ b/Sources/App.swift
-        @@ -1 +1,2 @@
-        +let title = "QuillCode"
-         import Foundation
-        """
-        let call = ToolCall(name: "host.git.diff", argumentsJSON: "{}")
-        let result = ToolResult(ok: true, stdout: diff)
-        let matchingComment = WorkspaceReviewCommentState(path: "Sources/App.swift", text: "Check the public API name.")
-        let lineComment = WorkspaceReviewCommentState(
-            path: "Sources/App.swift",
-            lineNumber: 1,
-            lineKind: .insertion,
-            text: "This line should stay public."
-        )
-        let rangeComment = WorkspaceReviewCommentState(
-            path: "Sources/App.swift",
-            lineNumber: 1,
-            endLineNumber: 2,
-            text: "Keep the title next to the import."
-        )
-        let staleComment = WorkspaceReviewCommentState(path: "README.md", text: "This file is no longer in the diff.")
-        let thread = ChatThread(
-            title: "Review changes",
-            events: [
-                ThreadEvent(kind: .toolQueued, summary: "host.git.diff queued", payloadJSON: try JSONHelpers.encodePretty(call)),
-                ThreadEvent(kind: .toolCompleted, summary: "host.git.diff completed", payloadJSON: try JSONHelpers.encodePretty(result)),
-                ThreadEvent(kind: .reviewComment, summary: "Commented on Sources/App.swift", payloadJSON: try JSONHelpers.encodePretty(matchingComment)),
-                ThreadEvent(kind: .reviewComment, summary: "Commented on Sources/App.swift:1", payloadJSON: try JSONHelpers.encodePretty(lineComment)),
-                ThreadEvent(kind: .reviewComment, summary: "Commented on Sources/App.swift:1-2", payloadJSON: try JSONHelpers.encodePretty(rangeComment)),
-                ThreadEvent(kind: .reviewComment, summary: "Commented on README.md", payloadJSON: try JSONHelpers.encodePretty(staleComment))
-            ]
-        )
-        let model = QuillCodeWorkspaceModel(root: QuillCodeRootState(
-            threads: [thread],
-            selectedThreadID: thread.id
-        ))
-
-        let review = model.surface().review
-
-        XCTAssertEqual(review.files.count, 1)
-        XCTAssertEqual(review.files.first?.comments.map(\.text), ["Check the public API name."])
-        XCTAssertEqual(
-            review.files.first?.hunkItems.first?.lines.first?.comments.map(\.text),
-            ["This line should stay public.", "Keep the title next to the import."]
-        )
-        XCTAssertEqual(review.files.first?.hunkItems.first?.lines.first?.comments.last?.lineRangeLabel, "Lines 1-2")
-    }
-
-    func testGitDiffReviewSurfaceHidesStaleDiffWhenLatestDiffFailed() throws {
-        let successfulCall = ToolCall(id: "git-diff-1", name: "host.git.diff", argumentsJSON: "{}")
-        let failedCall = ToolCall(id: "git-diff-2", name: "host.git.diff", argumentsJSON: "{}")
-        let successfulResult = ToolResult(ok: true, stdout: """
-        diff --git a/A.swift b/A.swift
-        --- a/A.swift
-        +++ b/A.swift
-        @@ -1 +1 @@
-        -old
-        +new
-        """)
-        let failedResult = ToolResult(ok: false, error: "not a git repository")
-        let thread = ChatThread(
-            title: "Git diff",
-            events: [
-                ThreadEvent(kind: .toolQueued, summary: "host.git.diff queued", payloadJSON: try JSONHelpers.encodePretty(successfulCall)),
-                ThreadEvent(kind: .toolCompleted, summary: "host.git.diff completed", payloadJSON: try JSONHelpers.encodePretty(successfulResult)),
-                ThreadEvent(kind: .toolQueued, summary: "host.git.diff queued", payloadJSON: try JSONHelpers.encodePretty(failedCall)),
-                ThreadEvent(kind: .toolFailed, summary: "host.git.diff failed", payloadJSON: try JSONHelpers.encodePretty(failedResult))
-            ]
-        )
-        let model = QuillCodeWorkspaceModel(root: QuillCodeRootState(
-            threads: [thread],
-            selectedThreadID: thread.id
-        ))
-
-        XCTAssertFalse(model.surface().review.isVisible)
-    }
-
 }
